requirements.txt
cassandra-driver==3.1.0
Django==1.9.8
django-cassandra-engine==0.7.0
futures==3.0.5
six==1.10.0
